+++ include/KSink.h
@@ -148,7 +148,7 @@ public:
 		return get_connection()->st.selector;
 	}
 	virtual void shutdown() = 0;
-	virtual uint8_t get_server_model() = 0;
+	virtual uint32_t get_server_model() = 0;
 	virtual KOPAQUE get_server_opaque() = 0;
 	/*
 	kserver* get_bind_server()
